WIP: processes rendering datagrid
