--- docs/content.example
@@ -1,1 +1,0 @@
-this-is-a-string [ geddy, neil, alex ] this-is-another-string
