+++ processes/rendering.ts
@@ -1,0 +1,8 @@
+import demo from '../modules/renderings/demo'
+import datagrid from '../modules/renderings/vue/datagrid'
+
+
+export default () => {
+  demo()
+  datagrid()
+}
